Filter rows of the melted similarity matrix to keep pairs with the same values in specific columns.

sim_all_same Filters melted similarity matrix to keep pairs with the same values in specific columns.

sim_all_same(
  sim_df,
  all_same_cols,
  annotation_cols = NULL,
  include_group_tag = FALSE,
  drop_lower = FALSE
)

Arguments

sim_df
tbl with melted similarity matrix.

all_same_cols
character vector specifying columns.

annotation_cols
optional character vector specifying which columns from metadata to annotate the left index of the filtered sim_df with.

include_group_tag
optional boolean specifying whether to include an identifier for the pairs using the values in the all_same_cols columns.

drop_lower
optional boolean specifying whether to drop the pairs where the first index is smaller than the second index. This is equivalent to dropping the lower triangular of sim_df.

Value

filtered sim_df where only pairs with the same values in all_same_cols columns are kept.

Examples

suppressMessages(suppressWarnings(library(magrittr)))
n <- 5
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), n, replace = TRUE),
  Metadata_type = sample(c("x", "y"), n, replace = TRUE),
  x = rnorm(n),
  y = x + rnorm(n) / 100,
  z = y + rnorm(n) / 1000
)
annotation_cols <- c("Metadata_group", "Metadata_type")
sim_df <- matric::sim_calculate(population, method = "pearson")
sim_df <- matric::sim_annotate(sim_df, annotation_cols)
all_same_cols <- c("Metadata_group")
include_group_tag <- TRUE
drop_lower <- FALSE
matric::sim_all_same(sim_df, all_same_cols, annotation_cols, include_group_tag, drop_lower)

#> # A tibble: 12 x 5
#>      id1   id2    sim Metadata_group Metadata_type
#>    <int> <int>  <dbl> <chr>          <chr>        
#>  1     1     2 -0.949 a              y            
#>  2     1     4  0.963 a              y            
#>  3     1     5 -0.916 a              y            
#>  4     2     1 -0.949 a              x            
#>  5     2     4 -0.999 a              x            
#>  6     2     5  0.996 a              x            
#>  7     4     1  0.963 a              x            
#>  8     4     2 -0.999 a              x            
#>  9     4     5 -0.990 a              x            
#> 10     5     1 -0.916 a              y            
#> 11     5     2  0.996 a              y            
#> 12     5     4 -0.990 a              y