Filter a melted similarity matrix to keep pairs with the same values in specific columns, and keep only some of these pairs.

sim_filter_all_same filters a melted similarity matrix to keep pairs with the same values in specific columns, keeping only some of these pairs.

sim_filter_all_same_keep_some(
  sim_df,
  row_metadata,
  all_same_cols,
  filter_keep_right,
  annotation_cols = NULL,
  drop_reference = TRUE,
  sim_cols = c("id1", "id2", "sim")
)

Arguments

sim_df
data.frame with melted similarity matrix.

row_metadata
data.frame with row metadata.

all_same_cols
character vector specifying columns.

filter_keep_right
data.frame of metadata specifying which rows to keep on the right index.

annotation_cols
optional character vector specifying which columns from metadata to annotate the left index of the filtered sim_df with.

drop_reference
optional boolean specifying whether to filter (drop) pairs using filter_keep_right on the left index.

sim_cols
optional character string specifying minimal set of columns for a similarity matrix

Value

Filtered sim_df as a data.frame, where only pairs with the same values in all_same_cols columns are kept, with further filtering using filter_keep_right.Rows are annotated based on the first index, if specified.

Examples

suppressMessages(suppressWarnings(library(magrittr)))
n <- 20
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), n, replace = TRUE),
  Metadata_type = sample(c("x", "y"), n, replace = TRUE),
  x = rnorm(n),
  y = x + rnorm(n) / 100,
  z = y + rnorm(n) / 1000
)
annotation_cols <- c("Metadata_group", "Metadata_type")
sim_df <- matric::sim_calculate(population, method = "pearson")
row_metadata <- attr(sim_df, "row_metadata")
sim_df <- matric::sim_annotate(sim_df, row_metadata, annotation_cols)
all_same_cols <- c("Metadata_group")
filter_keep_right <- tibble::tibble(Metadata_group = "a", Metadata_type = "x")
drop_reference <- FALSE
matric::sim_filter_all_same_keep_some(
  sim_df,
  row_metadata,
  all_same_cols,
  filter_keep_right,
  annotation_cols,
  drop_reference
)

#>    id1 id2        sim Metadata_group Metadata_type
#> 1    3   9  0.9999883              a             x
#> 2    3  13 -0.9898513              a             x
#> 3    3  14 -0.9976366              a             x
#> 4    3  15 -0.9996513              a             x
#> 5    3  16 -0.9997147              a             x
#> 6    3  19  0.9700913              a             x
#> 7    3  20  0.9999910              a             x
#> 8    9   3  0.9999883              a             x
#> 9    9  13 -0.9905275              a             x
#> 10   9  14 -0.9979574              a             x
#> 11   9  15 -0.9997674              a             x
#> 12   9  16 -0.9998186              a             x
#> 13   9  19  0.9712548              a             x
#> 14   9  20  0.9999998              a             x
#> 15  10   3 -0.9263145              a             y
#> 16  10   9 -0.9281271              a             y
#> 17  10  13  0.9704526              a             y
#> 18  10  14  0.9500125              a             y
#> 19  10  15  0.9359397              a             y
#> 20  10  16  0.9350490              a             y
#> 21  10  19 -0.9900624              a             y
#> 22  10  20 -0.9279062              a             y
#> 23  11   3 -0.8389305              a             y
#> 24  11   9 -0.8415548              a             y
#> 25  11  13  0.9077565              a             y
#> 26  11  14  0.8743433              a             y
#> 27  11  15  0.8530087              a             y
#> 28  11  16  0.8516904              a             y
#> 29  11  19 -0.9459479              a             y
#> 30  11  20 -0.8412342              a             y
#> 31  13   3 -0.9898513              a             x
#> 32  13   9 -0.9905275              a             x
#> 33  13  14  0.9972763              a             x
#> 34  13  15  0.9932585              a             x
#> 35  13  16  0.9929632              a             x
#> 36  13  19 -0.9947412              a             x
#> 37  13  20 -0.9904459              a             x
#> 38  14   3 -0.9976366              a             x
#> 39  14   9 -0.9979574              a             x
#> 40  14  13  0.9972763              a             x
#> 41  14  15  0.9991030              a             x
#> 42  14  16  0.9989931              a             x
#> 43  14  19 -0.9844776              a             x
#> 44  14  20 -0.9979194              a             x
#> 45  15   3 -0.9996513              a             x
#> 46  15   9 -0.9997674              a             x
#> 47  15  13  0.9932585              a             x
#> 48  15  14  0.9991030              a             x
#> 49  15  16  0.9999968              a             x
#> 50  15  19 -0.9761626              a             x
#> 51  15  20 -0.9997544              a             x
#> 52  16   3 -0.9997147              a             x
#> 53  16   9 -0.9998186              a             x
#> 54  16  13  0.9929632              a             x
#> 55  16  14  0.9989931              a             x
#> 56  16  15  0.9999968              a             x
#> 57  16  19 -0.9756124              a             x
#> 58  16  20 -0.9998071              a             x
#> 59  18   3 -0.9997947              a             y
#> 60  18   9 -0.9996850              a             y
#> 61  18  13  0.9867690              a             y
#> 62  18  14  0.9960397              a             y
#> 63  18  15  0.9989112              a             y
#> 64  18  16  0.9990256              a             y
#> 65  18  19 -0.9649742              a             y
#> 66  18  20 -0.9996997              a             y
#> 67  19   3  0.9700913              a             x
#> 68  19   9  0.9712548              a             x
#> 69  19  13 -0.9947412              a             x
#> 70  19  14 -0.9844776              a             x
#> 71  19  15 -0.9761626              a             x
#> 72  19  16 -0.9756124              a             x
#> 73  19  20  0.9711134              a             x
#> 74  20   3  0.9999910              a             x
#> 75  20   9  0.9999998              a             x
#> 76  20  13 -0.9904459              a             x
#> 77  20  14 -0.9979194              a             x
#> 78  20  15 -0.9997544              a             x
#> 79  20  16 -0.9998071              a             x
#> 80  20  19  0.9711134              a             x