Read similarity matrix.

sim_read reads similarity matrix.

sim_read(input, file_format = "parquet")

Arguments

input
character string specifying the input filename or directory.

file_format
character string specify file format. This must be one of csv or parquet(default).

Value

tbl with similarity matrix.

Details

With the CSV format, the row_metadata and metric_metadata attributes are read in and added as attributes to the sim_df tbl.

This is not required for Parquet because it the attributes are saved. For Parquet format, sim_read is identical to arrow::read_parquet

Examples

suppressMessages(suppressWarnings(library(magrittr)))
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), 4, replace = TRUE),
  x = rnorm(4),
  y = x + rnorm(4) / 100,
  z = y + rnorm(4) / 1000
)
sim_df <- sim_calculate(population, method = "pearson")
sim_df %>% sim_write("/tmp/test", file_format = "csv")

#> INFO [2021-03-25 18:14:07] Creating /tmp/test ...
#> INFO [2021-03-25 18:14:07] Writing /tmp/test/test.csv ...
#> INFO [2021-03-25 18:14:07] Writing /tmp/test/test_metadata.csv ...
#> INFO [2021-03-25 18:14:07] Writing /tmp/test/test_metadata.json ...

#> NULL

sim_df_csv <- sim_read("/tmp/test", file_format = "csv")

#> INFO [2021-03-25 18:14:07] Reading /tmp/test/test.csv ...

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   id1 = col_double(),
#>   id2 = col_double(),
#>   sim = col_double()
#> )

#> INFO [2021-03-25 18:14:08] Reading /tmp/test/test_metadata.csv ...

#> 
#> ── Column specification ────────────────────────────────────────────────────────
#> cols(
#>   id = col_double(),
#>   Metadata_group = col_character()
#> )

#> INFO [2021-03-25 18:14:08] Reading /tmp/test/test_metadata.json ...

sim_df %>% sim_write("/tmp/test.parquet")

#> INFO [2021-03-25 18:14:08] Writing /tmp/test.parquet ...

#> NULL

sim_df_parquet1 <- sim_read("/tmp/test.parquet")
sim_df %>% arrow::write_parquet("/tmp/test.parquet")
sim_df_parquet2 <- arrow::read_parquet("/tmp/test.parquet")
all(sim_df_parquet1 == sim_df_parquet2)

#> [1] TRUE

all(sim_df_parquet1 == sim_df_csv)

#> [1] TRUE