Filter a melted similarity matrix to keep pairs with the same values in specific columns, and keep only some of these pairs.

sim_filter_all_same filters a melted similarity matrix to keep pairs with the same values in specific columns, keeping only some of these pairs.

sim_filter_all_same_keep_some(
  sim_df,
  row_metadata,
  all_same_cols,
  filter_keep_right,
  annotation_cols = NULL,
  drop_reference = TRUE,
  sim_cols = c("id1", "id2", "sim")
)

Arguments

sim_df

data.frame with melted similarity matrix.

row_metadata

data.frame with row metadata.

all_same_cols

character vector specifying columns.

filter_keep_right

data.frame of metadata specifying which rows to keep on the right index.

annotation_cols

optional character vector specifying which columns from metadata to annotate the left index of the filtered sim_df with.

drop_reference

optional boolean specifying whether to filter (drop) pairs using filter_keep_right on the left index.

sim_cols

optional character string specifying minimal set of columns for a similarity matrix

Value

Filtered sim_df as a data.frame, where only pairs with the same values in all_same_cols columns are kept, with further filtering using filter_keep_right.Rows are annotated based on the first index, if specified.

Examples

suppressMessages(suppressWarnings(library(magrittr)))
n <- 20
population <- tibble::tibble(
  Metadata_group = sample(c("a", "b"), n, replace = TRUE),
  Metadata_type = sample(c("x", "y"), n, replace = TRUE),
  x = rnorm(n),
  y = x + rnorm(n) / 100,
  z = y + rnorm(n) / 1000
)
annotation_cols <- c("Metadata_group", "Metadata_type")
sim_df <- matric::sim_calculate(population, method = "pearson")
row_metadata <- attr(sim_df, "row_metadata")
sim_df <- matric::sim_annotate(sim_df, row_metadata, annotation_cols)
all_same_cols <- c("Metadata_group")
filter_keep_right <-
  tibble::tibble(Metadata_group = "a", Metadata_type = "x")
drop_reference <- FALSE
matric::sim_filter_all_same_keep_some(
  sim_df,
  row_metadata,
  all_same_cols,
  filter_keep_right,
  annotation_cols,
  drop_reference
)
#>     id1 id2         sim Metadata_group Metadata_type
#> 1     2   3  0.34097772              a             y
#> 2     4   3 -0.99612631              a             x
#> 3     5   3  0.99865250              a             x
#> 4     7   3 -0.99524137              a             x
#> 5     8   3 -0.95982375              a             x
#> 6     9   3 -0.99578087              a             y
#> 7    11   3 -0.99386287              a             y
#> 8    12   3  0.99887494              a             x
#> 9    13   3 -0.99177796              a             x
#> 10   14   3 -0.99175575              a             y
#> 11   17   3 -0.99972960              a             x
#> 12   19   3 -0.90346749              a             x
#> 13   20   3 -0.99314070              a             y
#> 14    2   4 -0.42232107              a             y
#> 15    3   4 -0.99612631              a             x
#> 16    5   4 -0.99022061              a             x
#> 17    7   4  0.98281781              a             x
#> 18    8   4  0.93143111              a             x
#> 19    9   4  0.98385443              a             y
#> 20   11   4  0.98028576              a             y
#> 21   12   4 -0.99917562              a             x
#> 22   13   4  0.99918908              a             x
#> 23   14   4  0.97664590              a             y
#> 24   17   4  0.99381216              a             x
#> 25   19   4  0.86227426              a             x
#> 26   20   4  0.97901186              a             y
#> 27    2   5  0.29173238              a             y
#> 28    3   5  0.99865250              a             x
#> 29    4   5 -0.99022061              a             x
#> 30    7   5 -0.99895704              a             x
#> 31    8   5 -0.97309257              a             x
#> 32    9   5 -0.99920118              a             y
#> 33   11   5 -0.99826432              a             y
#> 34   12   5  0.99506793              a             x
#> 35   13   5 -0.98380039              a             x
#> 36   14   5 -0.99706944              a             y
#> 37   17   5 -0.99958923              a             x
#> 38   19   5 -0.92449560              a             x
#> 39   20   5 -0.99787039              a             y
#> 40    2   7 -0.24775432              a             y
#> 41    3   7 -0.99524137              a             x
#> 42    4   7  0.98281781              a             x
#> 43    5   7 -0.99895704              a             x
#> 44    8   7  0.98259840              a             x
#> 45    9   7  0.99998374              a             y
#> 46   11   7  0.99991221              a             y
#> 47   12   7 -0.98950084              a             x
#> 48   13   7  0.97458897              a             x
#> 49   14   7  0.99952261              a             y
#> 50   17   7  0.99723810              a             x
#> 51   19   7  0.94093664              a             x
#> 52   20   7  0.99980796              a             y
#> 53    2   8 -0.06349113              a             y
#> 54    3   8 -0.95982375              a             x
#> 55    4   8  0.93143111              a             x
#> 56    5   8 -0.97309257              a             x
#> 57    7   8  0.98259840              a             x
#> 58    9   8  0.98152317              a             y
#> 59   11   8  0.98497332              a             y
#> 60   12   8 -0.94543706              a             x
#> 61   13   8  0.91602303              a             x
#> 62   14   8  0.98786801              a             y
#> 63   17   8  0.96608928              a             x
#> 64   19   8  0.98745213              a             x
#> 65   20   8  0.98604972              a             y
#> 66    2  12  0.38517432              a             y
#> 67    3  12  0.99887494              a             x
#> 68    4  12 -0.99917562              a             x
#> 69    5  12  0.99506793              a             x
#> 70    7  12 -0.98950084              a             x
#> 71    8  12 -0.94543706              a             x
#> 72    9  12 -0.99030895              a             y
#> 73   11  12 -0.98749890              a             y
#> 74   13  12 -0.99673079              a             x
#> 75   14  12 -0.98456316              a             y
#> 76   17  12 -0.99750210              a             x
#> 77   19  12 -0.88212321              a             x
#> 78   20  12 -0.98647849              a             y
#> 79    2  13 -0.45847570              a             y
#> 80    3  13 -0.99177796              a             x
#> 81    4  13  0.99918908              a             x
#> 82    5  13 -0.98380039              a             x
#> 83    7  13  0.97458897              a             x
#> 84    8  13  0.91602303              a             x
#> 85    9  13  0.97585055              a             y
#> 86   11  13  0.97153528              a             y
#> 87   12  13 -0.99673079              a             x
#> 88   14  13  0.96720300              a             y
#> 89   17  13  0.98853399              a             x
#> 90   19  13  0.84118369              a             x
#> 91   20  13  0.97001203              a             y
#> 92    2  17 -0.31902541              a             y
#> 93    3  17 -0.99972960              a             x
#> 94    4  17  0.99381216              a             x
#> 95    5  17 -0.99958923              a             x
#> 96    7  17  0.99723810              a             x
#> 97    8  17  0.96608928              a             x
#> 98    9  17  0.99764544              a             y
#> 99   11  17  0.99616642              a             y
#> 100  12  17 -0.99750210              a             x
#> 101  13  17  0.98853399              a             x
#> 102  14  17  0.99446736              a             y
#> 103  19  17  0.91319103              a             x
#> 104  20  17  0.99559109              a             y
#> 105   2  19  0.09490553              a             y
#> 106   3  19 -0.90346749              a             x
#> 107   4  19  0.86227426              a             x
#> 108   5  19 -0.92449560              a             x
#> 109   7  19  0.94093664              a             x
#> 110   8  19  0.98745213              a             x
#> 111   9  19  0.93899049              a             y
#> 112  11  19  0.94534044              a             y
#> 113  12  19 -0.88212321              a             x
#> 114  13  19  0.84118369              a             x
#> 115  14  19  0.95094827              a             y
#> 116  17  19  0.91319103              a             x
#> 117  20  19  0.94739119              a             y